+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KQ8CA4FD5K7MFFRHCYGE0QTX"
+    id: Optional[str] = "01KQ9J5A497MS19M0DWPYYWED4"
 
     description: Optional[str] = None
 
@@ -1357,7 +1357,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KQ8CA4F8T2TB9AT8Q8NPKK9E"
+        "tool_01KQ9J5A3WMG3A19AAPBHJH3D7"
     )
 
     display_name: Optional[str] = None
@@ -1479,7 +1479,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KQ8CA4F7PVV3RPD3SK5XYQXB"
+    id: Optional[str] = "01KQ9J5A3WP3W9BEGB5PW1R1E6"
 
     description: Optional[str] = None
 
@@ -1623,7 +1623,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KQ8CA4F6GHREWC0AKDR3JSZ7"
+        "tool_01KQ9J5A3V4Y1BRYTKGZEWPFS8"
     )
 
     display_name: Optional[str] = None
@@ -1914,7 +1914,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KQ8CA4F5K3NZF7QK503A5E6G"
+        "tool_01KQ9J5A3SY6GNE6K9GRDESQ4R"
     )
 
     display_name: Optional[str] = None
@@ -2099,7 +2099,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KQ8CA4F4E93A51YJJE253N65"
+        "tool_01KQ9J5A3QZKGKGZX9SMW172FB"
     )
 
     display_name: Optional[str] = None
@@ -2288,7 +2288,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KQ8CA4F3JDTN6S44EYPS1S06"
+        "tool_01KQ9J5A3PASZNY89R61J6NKRM"
     )
 
     display_name: Optional[str] = None