+++ packages/orq-rc/RELEASES.md
@@ -7369,3 +7369,13 @@ Based on:
 - [python v4.9.0-rc.4] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.4] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.4 - packages/orq-rc
+
+## 2026-04-28 08:13:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.5] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.5] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.5 - packages/orq-rc